localization/requirements.txt
numpy==1.26
scipy==1.11
joblib==1.0.1
# numpy==1.21.0
pandas==1.3.0
paho-mqtt==1.5.1
scikit-learn==0.24.2
tensorflow==2.5.0